Add all modules(product/cart/category/order/orderItem/cartItem/notification/payment) and test controllers files

+++ src/main/java/com/retail_project/customer/CustomerService.java
@@ -1,40 +1,78 @@
 package com.retail_project.customer;
 
+import com.retail_project.exceptions.CustomerNotFoundException;
 import lombok.RequiredArgsConstructor;
+import org.apache.commons.lang3.StringUtils;
 import org.apache.kafka.common.message.LeaderAndIsrRequestData;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
 
+import static java.lang.String.format;
+
 @Service
 @RequiredArgsConstructor
 public class CustomerService {
     private final CustomerRepository customerRepository;
+    private final CustomerMapper mapper;
+
+    //create customer
+    public CustomerResponse createCustomer(CustomerRequest request){
+        var customer =mapper.toEntity(request);
+        return mapper.toResponse(customerRepository.save(customer));
+    }
 
     //get customer by id
-    public Optional<Customer> getCustomerById(int id){
-        return customerRepository.findById(id);
+    public CustomerResponse getCustomerById(int id){
+        return customerRepository.findById(id).map(mapper::toResponse).orElseThrow(()->new CustomerNotFoundException(id));
     }
     // get all customers
-    public List<Customer> getCustomers(){
-        return customerRepository.findAll();
+    public List<CustomerResponse> getCustomers(){
+        return customerRepository.findAll().stream().map(mapper::toResponse).toList();
     }
-    //create customer
-    public Customer createCustomer(Customer customer ){
-        return customerRepository.save(customer);
-    }
+
+
+
     // update customer by id
-    public Customer updateCustomer(int id,Customer customerDetails){
-       Customer customer= customerRepository.findById(id).orElseThrow();
-        customer.setFirstname(customerDetails.firstname);
-        customer.setLastname(customerDetails.lastname);
-        customer.setEmail(customerDetails.email);
-        customer.setAddress(customer.address);
-        return customer;
+    public CustomerResponse updateCustomer(int id,CustomerRequest request){
+        Customer customer =customerRepository.findById(id).orElseThrow(()->new CustomerNotFoundException(id));
+        customer.setFirstname(request.firstname());
+        customer.setLastname(request.lastname());
+        customer.setEmail(request.email());
+        customer.setAddress(request.address());
+        customerRepository.save(customer);
+        return mapper.toResponse(customer);
+    }
+
+    private void mergeCustomer(Customer customer, CustomerRequest request) {
+        if (StringUtils.isNotBlank(request.firstname())) {
+            customer.setFirstname(request.firstname());
+        }
+        if (StringUtils.isNotBlank(request.lastname())) {
+            customer.setLastname(request.lastname());
+        }
+        if (StringUtils.isNotBlank(request.email())) {
+            customer.setEmail(request.email());
+        }
+        if (request.address() != null) {
+            customer.setAddress(request.address());
+        }
     }
     // delete customer by id
     public void deleteCustomer (int id){
-         customerRepository.deleteById(id);
+        if (!customerRepository.existsById(id)) {
+            throw new RuntimeException("Customer not found");
+        }
+        customerRepository.deleteById(id);
+    }
+
+    //Pagination
+    public Page<Customer> getAllCustomer(int page,int size) {
+        Pageable pageable= PageRequest.of(page, size);
+        return customerRepository.findAll(pageable);
     }
 }
